+++ src/scripts/users.ts
@@ -5,6 +5,7 @@ import { UserFunctions, Protocol } from '../models';
 module.exports = function (app: Express) {
 	let userRepo = new UserRepository()
 
+	// Authenticate User
 	app.get('/api/authenticate', function (req: Request, res: Response, next: NextFunction) {
 		let userName = req.query.user_name
 		let userPassword = req.query.user_password
@@ -15,18 +16,22 @@ module.exports = function (app: Express) {
 
 		userRepo.findUser(userName).then((user) => {
 			if (UserFunctions.validatePassword(user, userPassword)) {
-				let publicUser = UserFunctions.toPublicUser(user)
-				Protocol.success(res, publicUser)
+				Protocol.createUserSession(req, user).then(() => {
+					let publicUser = UserFunctions.toPublicUser(user)
+					Protocol.success(res, publicUser)
+				}).catch(error => Protocol.error(res, "SESSION_CREATE_FAIL"))
 			} else {
 				Protocol.error(res, "USER_AUTH_FAIL", "Invalid Password")
 			}
 		}).catch(error => Protocol.error(res, "USER_QUERY_FAIL", "Could Not Find User"))
 	})
 
+	// Find User
 	app.get('/api/users/query', function (req: Request, res: Response, next: NextFunction) {
 		Protocol.error(res, "NOT_IMPLEMENTED", "Not yet implemented")
 	})
 
+	// Get User
 	app.get('/api/users', function (req: Request, res: Response, next: NextFunction) {
 		let id = req.query.user_id
 
@@ -40,6 +45,7 @@ module.exports = function (app: Express) {
 		}).catch(error => Protocol.error(res, "USER_QUERY_FAIL", "Could Not Find User"))
 	})
 
+	// Create user
 	app.post('/api/users', function (req: Request, res: Response, next: NextFunction) {
 		let userName = req.query.user_name
 		let userPassword = req.query.user_password
@@ -54,11 +60,16 @@ module.exports = function (app: Express) {
 		}).catch(error => Protocol.error(res, "USER_CREATE_FAIL", "Failed to create user"))
 	})
 
+	// Delete User
 	app.delete('/api/users', function (req: Request, res: Response, next: NextFunction) {
 		let user_id = req.query.user_id
 
 		if (!Protocol.validateParams([user_id])) {
 			return Protocol.error(res, "INVALID_PARAM")
+		}
+
+		if (!Protocol.validateUserSession(req, user_id)) {
+			return Protocol.error(res, "INSUFFICIENT_PERMISSIONS")
 		}
 
 		userRepo.delete(user_id).then((result) => {
